Change off-chain TXs to avoid previous attack
Previously it was possible for a number of colluding hops to illegally
consume the guests' virtual output with a TX_both for which the other
two inputs were fabricated, leaving the guests with no valid output to
fund their channel. To avoid this, the guests' public keys have been
added to all intermediary outputs that can be legally spent by TX_both.