Simplify F_Chan by allowing S to control values
The functionality now has correct behaviour when one party is corrupted.
On the downside, F_Chan is not informed about the opening, paying and
funding coins by the Environment, but by the Simulator. It is debatable
whether this expresses the intended security guarantees. The Simulator
and Lemmas have been changed respectively.