Prove base case of real balance security Lemma
Change proof approach: We now do induction on the payments and fundings
simultaneously. Base case is an on-chain channel that closes right away
securely, induction step is to either open a virtual on top or to do a
payment securely. The history of the "enabler" channel is taken into
account if we are a virtual. This way we cover all possible channel
histories from open to close. This approach avoids the circularity of
the old one.